Reorganized import and changed RI type to float

+++ RIAssigner/compute/ComputationMethod.py
@@ -1,10 +1,10 @@
 from abc import ABC, abstractmethod
-from RIAssigner.data import Data
 from typing import List
+from RIAssigner.data import Data
 
 
 class ComputationMethod(ABC):
 
     @abstractmethod
-    def compute(self, query: Data, reference: Data) -> List[int]:
+    def compute(self, query: Data, reference: Data) -> List[float]:
         ...
